t(): renamed folder, functions
