+++ src/test/java/functions/ClickAodocsWebElement.java
@@ -1,4 +1,4 @@
-package methods;
+package functions;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
