+++ src/test/java/functions/VisitWebpageDirect.java
@@ -1,4 +1,4 @@
-package methods;
+package functions;
 
 import org.junit.jupiter.api.Assertions;
 import org.openqa.selenium.WebDriver;
